+++ tests/s6/test.sh
@@ -22,7 +22,7 @@ cd top
 orbit lock --force
 
 # verify the ip dependency graph only has 1 aka version
-STDOUT=$(orbit tree --ip)
+STDOUT=$(orbit tree -e ip)
 
 cd ..